PARCHE CIERRE ENTREGA ENTRE ORGS

+++ InventarioREST32-master/src/main/java/com/bave/inventariorest/model/RcvTransactions.java
@@ -56,7 +56,7 @@ public class RcvTransactions {
     @Column(name = "CURRENCY_CONVERSION_TYPE")
     private String currencyConversionType;
     @Column(name = "CURRENCY_CONVERSION_RATE")
-    private Long currencyConversionRate;
+    private Double currencyConversionRate;
     @Column(name = "CURRENCY_CONVERSION_DATE")
     private String currencyConversionDate;
     @Column(name = "DELIVER_TO_LOCATION_ID")
@@ -266,11 +266,11 @@ public class RcvTransactions {
         this.currencyConversionType = currencyConversionType;
     }
 
-    public Long getCurrencyConversionRate() {
+    public Double getCurrencyConversionRate() {
         return currencyConversionRate;
     }
 
-    public void setCurrencyConversionRate(Long currencyConversionRate) {
+    public void setCurrencyConversionRate(Double currencyConversionRate) {
         this.currencyConversionRate = currencyConversionRate;
     }
